[issue#47] Add Feedback aggregate before adding karma test
