+++ src/main/java/com/vsct/dt/hesperides/feedback/FeedbackConfiguration.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.vsct.dt.hesperides.resources;
+package com.vsct.dt.hesperides.feedback;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.hibernate.validator.constraints.NotEmpty;
