+++ src/main/java/com/vsct/dt/hesperides/feedback/Feedbacks.java
@@ -1,0 +1,42 @@
+/*
+ *
+ *  * This file is part of the Hesperides distribution.
+ *  * (https://github.com/voyages-sncf-technologies/hesperides)
+ *  * Copyright (c) 2016 VSCT.
+ *  *
+ *  * Hesperides is free software: you can redistribute it and/or modify
+ *  * it under the terms of the GNU General Public License as
+ *  * published by the Free Software Foundation, version 3.
+ *  *
+ *  * Hesperides is distributed in the hope that it will be useful, but
+ *  * WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ *  * General Public License for more details.
+ *  *
+ *  * You should have received a copy of the GNU General Public License
+ *  * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ *
+ *
+ */
+
+package com.vsct.dt.hesperides.feedback;
+
+import com.vsct.dt.hesperides.feedback.jsonObject.FeedbackJson;
+import com.vsct.dt.hesperides.security.model.User;
+
+/**
+ *  Feedback aggragator interface.
+ *
+ * Created by stephane_fret on 07/03/2017.
+ */
+public interface Feedbacks {
+    /**
+     * Send a feedback to Hipchat
+     *
+     * @param user : the current user
+     * @param feedbackJson     : feedback message with screenshoot and url
+     */
+    void sendFeedbackToHipchat(final User user,
+                               final FeedbackJson feedbackJson);
+}
